% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_att.R
\name{nc_atts}
\alias{nc_atts}
\alias{nc_atts}
\alias{nc_atts.NetCDF}
\alias{nc_atts}
\alias{nc_atts.character}
\title{NetCDF attributes}
\usage{
nc_atts(x, ...)

\method{nc_atts}{NetCDF}(x, ...)

\method{nc_atts}{character}(x, ...)
}
\arguments{
\item{x}{filename or handle}

\item{...}{ignored}
}
\description{
All attributes in the file, globals are treated as if they bleong to variable 'NC_GLOBAL'.
}
