% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_axis.R
\name{nc_axis}
\alias{nc_axis}
\alias{nc_axis.character}
\alias{nc_axis.NetCDF}
\title{NetCDF axes}
\usage{
nc_axis(x, i)

\method{nc_axis}{character}(x, i)

\method{nc_axis}{NetCDF}(x, i)
}
\arguments{
\item{x}{NetCDF source}

\item{i}{index of axis (1-based, 0 is "empty")}
}
\value{
data frame of a single axis instance, with columns 'axis', 'variable', 'dimension'
}
\description{
An \code{axis} is an instance of a dimension.
}
\details{
Each data source has a set of dimensions available for use by variables. Each axis is
a 1-dimensional instance.
}
\examples{
f <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncmeta")
nc_axis(f, 1)
}
