% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_coords.R
\name{nc_coords}
\alias{nc_coords}
\alias{nc_coords.character}
\alias{nc_coords.NetCDF}
\title{NetCDF Coordinates}
\usage{
nc_coords(x, ...)

\method{nc_coords}{character}(x, ...)

\method{nc_coords}{NetCDF}(x, ...)
}
\arguments{
\item{x}{NetCDF source}

\item{...}{ignored}
}
\value{
tibble with "variable", "X", "Y", "Z", "T", and "order" columns. The "variable",
"X", "Y", "Z", and "T" columns of the response reference variables by name.
}
\description{
In NetCDF, variables are defined along dimensions and are said to have "coordinate
variables" that define the typically spatio-temporal positions of the data's cells.
}
\details{
This function attempts to identify the X, Y, Z, and T coordinate variables for each
data variable in the provided NetCDF source. The NetCDF-CF attribute conventions are
used to make this determination.

See \url{http://cfconventions.org/cf-conventions/v1.6.0/cf-conventions.html#coordinate-system}
for more.
}
