% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_dim.R
\name{nc_dim}
\alias{nc_dim}
\alias{nc_dim.character}
\alias{nc_dim.NetCDF}
\alias{nc_dim.ncdf4}
\title{NetCDF dimension}
\usage{
nc_dim(x, i, ...)

\method{nc_dim}{character}(x, i, ...)

\method{nc_dim}{NetCDF}(x, i, ...)

\method{nc_dim}{ncdf4}(x, i, ...)
}
\arguments{
\item{x}{filename or handle}

\item{i}{index of dimension (zero based)}

\item{...}{ignored}
}
\value{
data frame of dimension information, with columns 'id', 'name', 'length', 'unlim'
}
\description{
Obtain information about a single dimension by index.
}
\examples{
f <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncmeta")
nc_dim(f, 0)
}
\seealso{
\code{nc_dims} to obtain information about all dimensions, \code{nc_inq} for an
overview of the file
}
