% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_dimension.R
\name{nc_dims}
\alias{nc_dims}
\alias{nc_dims.character}
\alias{nc_dims.NetCDF}
\alias{nc_dims.ncdf4}
\title{NetCDF dimension}
\usage{
nc_dims(x, ...)

\method{nc_dims}{character}(x, ...)

\method{nc_dims}{NetCDF}(x, ...)

\method{nc_dims}{ncdf4}(x, ...)
}
\arguments{
\item{x}{file address or handle}

\item{...}{ignored}
}
\value{
data frame of dimension information, one row per dimension, with columns
'id', 'name', 'length', 'unlim'
}
\description{
Get information about the dimensions in a NetCDF source.
}
\examples{
f <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncmeta")
nc_dims(f)
}
