% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_extended.R
\name{nc_extended}
\alias{nc_extended}
\alias{nc_extended.character}
\alias{nc_extended.NetCDF}
\alias{nc_extended.ncdf4}
\title{NetCDF extended dimension attributes}
\usage{
nc_extended(x, ...)

\method{nc_extended}{character}(x, ...)

\method{nc_extended}{NetCDF}(x, ...)

\method{nc_extended}{ncdf4}(x, ...)
}
\arguments{
\item{x}{filename or handle}

\item{...}{ignored currently}
}
\value{
data frame of extended dimension attribute information
}
\description{
Generate a table of all extended dimension attributes. For now that means
interpretation of any "time" dimension.
}
\examples{
f <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncmeta")
nc_extended(f)

}
