% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nc_grid_mapping}
\alias{nc_grid_mapping}
\alias{nc_grid_mapping.character}
\alias{nc_grid_mapping.NetCDF}
\alias{nc_grid_mapping.data.frame}
\title{Get Grid Mapping}
\usage{
nc_grid_mapping(x, ...)

\method{nc_grid_mapping}{character}(x, ...)

\method{nc_grid_mapping}{NetCDF}(x, ...)

\method{nc_grid_mapping}{data.frame}(x, ...)
}
\arguments{
\item{x}{open NetCDF object, character file path or url to be
opened with RNetCDF::open.nc, or data.frame as returned from ncmeta::nc_atts}
}
\value{
tibble containing attributes that make up the file's grid_mapping
}
\description{
Get the grid mapping from a NetCDF file
}
