% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_sources.R
\name{nc_sources}
\alias{nc_sources}
\alias{nc_sources.character}
\title{NetCDF sources}
\usage{
nc_sources(x, ...)

\method{nc_sources}{character}(x, ...)
}
\arguments{
\item{x}{data source string}

\item{...}{ignored}
}
\value{
data frame with columns 'access' (time of access), 'source' (the source string)
}
\description{
A record of file, URL, or any data source with NetCDF.
}
\examples{
f <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncmeta")
nc_sources(f)
}
