% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_variable.R
\name{nc_vars}
\alias{nc_vars}
\alias{nc_vars.character}
\alias{nc_vars.NetCDF}
\title{NetCDF variables}
\usage{
nc_vars(x, ...)

\method{nc_vars}{character}(x, ...)

\method{nc_vars}{NetCDF}(x, ...)
}
\arguments{
\item{x}{filename or handle}

\item{...}{ignored currently}
}
\value{
data frame of variable information
}
\description{
Generate a table of all variables.
}
\examples{
f <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncmeta")
nc_vars(f)

}
