% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_nc_connection.R
\name{nc_connection}
\alias{nc_connection}
\title{NetCDF connection}
\usage{
nc_connection(x, preference = NULL, ...)
}
\arguments{
\item{x}{character string source}

\item{preference}{preferred engine}

\item{...}{ignored currently}
}
\value{
NetCDF connection object
}
\description{
Currently ncdf4 or RNetCDF are available as preference, with default RNetCDF.
}
