% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_dim.R
\name{nc_dim}
\alias{nc_dim}
\alias{nc_dim}
\alias{nc_dim.character}
\alias{nc_dim}
\alias{nc_dim.NetCDF}
\alias{nc_dim}
\alias{nc_dim.ncdf4}
\title{NetCDF variables
Obtain information about a single dimension by index.}
\usage{
nc_dim(x, i, ...)

\method{nc_dim}{character}(x, i, ...)

\method{nc_dim}{NetCDF}(x, i, ...)

\method{nc_dim}{ncdf4}(x, i, ...)
}
\arguments{
\item{x}{filename or handle}

\item{i}{index of dimension (zero based)}

\item{...}{ignored}
}
\description{
NetCDF variables
Obtain information about a single dimension by index.
}
\seealso{
\code{nc_vars`` to obtain information about all dimensions,}nc_inq` for an
overview of the file
}
