% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_dimension.R
\name{nc_dims}
\alias{nc_dims}
\alias{nc_dims.character}
\alias{nc_dims.default}
\title{NetCDF dimension}
\usage{
nc_dims(x, ...)

\method{nc_dims}{character}(x, ...)

\method{nc_dims}{default}(x, ...)
}
\arguments{
\item{x}{file address or handle}

\item{...}{ignored}
}
\description{
Get information about the dimensions in a NetCDF source.
}
