% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_extended.R
\name{nc_extended}
\alias{nc_extended}
\alias{nc_extended.character}
\alias{nc_extended.NetCDF}
\title{NetCDF extended dimension attributes}
\usage{
nc_extended(x, ...)

\method{nc_extended}{character}(x, ...)

\method{nc_extended}{NetCDF}(x, ...)
}
\arguments{
\item{x}{filename or handle}

\item{...}{ignored currently}
}
\value{
data frame of extended dimension attribute information
}
\description{
Generate a table of all extended dimension attributes. For now that means
interpretation of any "time" dimension.
}
