% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_grid.R
\name{nc_grids}
\alias{nc_grids}
\alias{nc_grids}
\alias{nc_grids.character}
\alias{nc_grids}
\alias{nc_grids.NetCDF}
\title{NetCDF grids}
\usage{
nc_grids(x, ...)

\method{nc_grids}{character}(x, ...)

\method{nc_grids}{NetCDF}(x, ...)
}
\arguments{
\item{x}{NetCDF source}

\item{...}{ignored}
}
\description{
An \code{grid} is a discretized space, defined by a set of dimensions.
}
\details{
Each data source has a set of dimensions available for use by variables. Each grid is
an n-dimensional space available for use by 0, 1 or more variables. A grid only really
exists if  variable is defined for it, and 'grid' is an implicit entity not an explicit
part of the NetCDF API definition. The Unidata pages refer to "shape", which is more or less what
we mean by "grid".
@name nc_grids
}
