% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_var.R
\name{nc_var}
\alias{nc_var}
\alias{nc_var}
\alias{nc_var.character}
\alias{nc_var}
\alias{nc_var.NetCDF}
\title{NetCDF variable}
\usage{
nc_var(x, i)

\method{nc_var}{character}(x, i)

\method{nc_var}{NetCDF}(x, i)
}
\arguments{
\item{x}{file name or handle}

\item{i}{variable index (zero based)}
}
\value{
data frame of variable information
}
\description{
Return a data frame about the variable at index \code{i}.
}
\seealso{
\code{nc_vars`` to obtain information about all variables,}nc_inq` for an
overview of the file
}
