% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_variable.R
\name{nc_vars}
\alias{nc_vars}
\alias{nc_vars.character}
\alias{nc_vars.NetCDF}
\alias{nc_vars.ZarrGroup}
\title{NetCDF variables}
\usage{
nc_vars(x, ...)

\method{nc_vars}{character}(x, ...)

\method{nc_vars}{NetCDF}(x, ...)

\method{nc_vars}{ZarrGroup}(x, ...)
}
\arguments{
\item{x}{filename or handle}

\item{...}{ignored currently}
}
\value{
data frame of variable information
}
\description{
Generate a table of all variables.
}
